# [Metal] Use managed storage mode to improve performance