It turned out the current `.shared` storage mode has made things pretty slow. This PR switched to use the `.managed` storage mode for `MTLBuffer`s, see https://developer.apple.com/documentation/metal/setting_resource_storage_modes/choosing_a_resource_storage_mode_in_macos.

Benchmark result:

| example       | old FPS (.shared) | new FPS (.managed) |
|---------------|-------------------|--------------------|
| mpm88         | 21-22             | 25-28              |
| mpm_lag_force | 11                | 13                 |
| stable_fluid  | 23                | 50                 |
| vortex*        | 1.2               | 6                  |

* I shamelessly copied the vortex example from here https://forum.taichi.graphics/t/vortex-method-demo/775. Note that the original example's `fix_streamfunc` had a global reduction to a kernel temporary variable. This wasn't supported in TLS yet, so I changed that to use a `ti.var` to store the global sum result. Also note that the global reduction here hurt the performance so much that, if I reduce the #threads from 262144 to 65536 (i.e. grid-stride loop), because of TLS and the reduced number of fatomic add, the FPS went to 12.
* The main bottleneck of mpm_lag_force is now GUI (30% of the time on GUI, only 12% on Metal. 4.4% out of that 12% was spent on `Snode::write_float` )

Related issue = #935

[[Click here for the format server]](http://kun.csail.mit.edu:31415/)

----