@k-ye 2020-07-05
@yuanming-hu not related to this PR. I found that just by enabling fast math, `sdf_render.py` went from 25fps to 41fps. 

* Do you know if fast math is always on on CUDA?
* If not, shall we add this to the compile config, e.g. `ti.init(..., use_fast_math=True)`?