@archibate 2020-07-05
> ti.init(..., fast_math=True)

#1396, should I add this option to that examp?