@k-ye 2020-07-05
> #1396, should I add this option to that examp?

Sure! Sorry i'll have to take a look tmr. But i see that you've already added `fast_math`..?