@yuanming-hu 2020-07-06
By default, `CompileConfig::fast_math` is on:

https://github.com/taichi-dev/taichi/blob/571ff554d752300bc2a0b756b5e4b18b3df16e4e/taichi/program/compile_config.cpp#L35

> * Do you know if fast math is always on on CUDA?

It follows ``CompileConfig::fast_math`:

https://github.com/taichi-dev/taichi/blob/571ff554d752300bc2a0b756b5e4b18b3df16e4e/taichi/backends/cuda/jit_cuda.cpp#L174

> * If not, shall we add this to the compile config, e.g. `ti.init(..., use_fast_math=True)`? (fast math broke some examples on Metal, so I don't think it can be always on..)

I believe `ti.init(fast_math=True/False)` is already supported - which Metal example does it break?